Move item preparation to SolrQueryBaseService.
